rndartt: spell the 64-bit palette and row arithmetic as long long
FindNearestEntry computes its distances with dsubu, daddu, and a
64-bit multiply, and the BMP, canvas, and GIF row offsets multiply
through CxxMultiplyInt64 before narrowing, so those values and casts
now use long long instead of a long that is four bytes wide on
current toolchains.

+++ src/rndartt/abmpfile.cpp
@@ -391,8 +391,8 @@ int ABmpFile::ReadRlePixels(ABitmap *pImage) {
                 pDest += abDelta[0];
                 nColumn += abDelta[0];
                 y += abDelta[1] * mRowStep;
-                pRow += static_cast<long>(abDelta[1]) * nStride;
-                pDest += static_cast<long>(abDelta[1]) * nStride;
+                pRow += static_cast<long long>(abDelta[1]) * nStride;
+                pDest += static_cast<long long>(abDelta[1]) * nStride;
             } else if (mBitCount == kBitCount8) {
                 fread(pDest, 1, abPair[1], mFile);
                 pDest += abPair[1];
